+++ changelog.md
@@ -8,6 +8,7 @@
 - Cleaned up outlines on Blank Checks sprite
 - Cleaned up outlines on all of Bubblebeam's sprites
 - Fixed not being able to aim Ki Blast's mastery projectile when on controller
+- Damage from Ki Blast's mastery projectile now scales with the current floor's enemy health modifier
 
 ## 1.26.2 (2025-06-06)